requirements.txt
dacite
PyYAML
aiopg
psycopg2-binary
peewee
peewee-async
peewee-migrate2
oitnb
click
pytest-asyncio
fastapi
uvicorn
httpx
arq
testing.redis
